+++ src/DependencyInjection/Configuration.php
@@ -30,6 +30,10 @@ class Configuration implements ConfigurationInterface
                     ->info('If true, that backend lost password link will be automatically added to the backed login template. Default false. Will be true in the next major version!')
                     ->defaultFalse()
                 ->end()
+                ->integerNode('nc_notification')
+                    ->info('The numeric ID of the notification center notification which is sent for resetting the password.')
+                    ->defaultFalse()
+                ->end()
             ->end();
 
         return $treeBuilder;