+++ src/Resources/contao/languages/de/default.php
@@ -1,15 +1,21 @@
 <?php
 
+/*
+ * Copyright (c) 2021 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
+ */
+
 $lang = &$GLOBALS['TL_LANG']['MSC']['backendLostPassword'];
 
 /**
- * Misc
+ * Misc.
  */
 $lang = [
     'thankYou' => 'Vielen Dank',
     'request' => 'Neues Passwort anfordern',
-    'requestExplanationEmail' => 'Bitte geben Sie hier die E-Mail-Adresse Ihres Contao-Benutzers ein:',
-    'requestLinkSentEmail' => 'Wenn es einen Contao-Benutzer mit der von Ihnen eingegebenen E-Mail-Adresse gibt, erhalten Sie in Kürze eine E-Mail mit allen weiteren Informationen zum Zurücksetzen Ihres Passworts.',
+    'requestExplanationEmail' => 'Bitte geben Sie hier die E-Mail-Adresse oder den Benutzernamen Ihres Contao-Benutzers ein:',
+    'requestLinkSentEmail' => 'Wenn es einen Contao-Benutzer mit der von Ihnen eingegebenen E-Mail-Adresse bzw. Benutzernamen gibt, erhalten Sie in Kürze eine E-Mail mit allen weiteren Informationen zum Zurücksetzen Ihres Passworts.',
     'spamNote' => '<strong>WICHTIG:</strong> Sollten Sie keine E-Mail erhalten, sehen Sie bitte im SPAM-Ordner Ihres E-Mail-Kontos nach.',
     'messageSubjectResetPassword' => 'Passwort zurücksetzen',
     'messageBodyResetPassword' => 'Sehr geehrte Damen und Herren,
@@ -22,5 +28,5 @@ Klicken Sie auf den folgenden Link, um ein neues Passwort zu vergeben:
 
 Vielen Dank.',
     'reset' => 'Passwort zurücksetzen',
-    'resetErrorExplanation' => 'Der Link zum Zurücksetzen Ihres Passworts, den Sie gerade aufgerufen haben, ist ungültig oder wurde bereits aufgerufen.'
+    'resetErrorExplanation' => 'Der Link zum Zurücksetzen Ihres Passworts, den Sie gerade aufgerufen haben, ist ungültig oder wurde bereits aufgerufen.',
 ];