+++ src/Resources/contao/languages/de/tl_nc_notification.php
@@ -1,0 +1,14 @@
+<?php
+
+/*
+ * Copyright (c) 2021 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
+ */
+
+$lang = &$GLOBALS['TL_LANG']['tl_nc_notification'];
+
+/*
+ * Types
+ */
+$lang['type']['user_password'][0] = 'Benutzer: Passwort vergessen';