+++ src/Resources/contao/languages/en/default.php
@@ -1,15 +1,21 @@
 <?php
 
+/*
+ * Copyright (c) 2021 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
+ */
+
 $lang = &$GLOBALS['TL_LANG']['MSC']['backendLostPassword'];
 
 /**
- * Misc
+ * Misc.
  */
 $lang = [
     'thankYou' => 'Thank you',
     'request' => 'Request new password',
-    'requestExplanationEmail' => 'Please type in the email address of your Contao user here:',
-    'requestLinkSentEmail' => 'Wenn es einen Contao-Benutzer mit dieser E-Mail-Adresse gibt, erhalten Sie in Kürze eine E-Mail mit allen weiteren Informationen zum Zurücksetzen Ihres Passworts.',
+    'requestExplanationEmail' => 'Please type in the email address or username of your Contao user here:',
+    'requestLinkSentEmail' => 'If there\'s a contao user with the email address or username you typed in, you\'ll get an email with the further steps in no time.',
     'spamNote' => '<strong>IMPORTANT:</strong> If you don\'t receive an email, please take a look in the SPAM folder of your email account.',
     'messageSubjectResetPassword' => 'Reset password',
     'messageBodyResetPassword' => 'Dear Sir or Madam,
@@ -22,5 +28,5 @@ Click the following link to set a new password:
 
 Thank you.',
     'reset' => 'Reset password',
-    'resetErrorExplanation' => 'The link for resetting your password you just opened, is invalid or has already been used.'
+    'resetErrorExplanation' => 'The link for resetting your password you just opened, is invalid or has already been used.',
 ];